+++ cfn-resources/project-ip-access-list/cmd/resource/list.go
@@ -20,17 +20,15 @@ import (
 
 	"github.com/aws-cloudformation/cloudformation-cli-go-plugin/cfn/handler"
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/cloudformation"
 	"github.com/mongodb/mongodbatlas-cloudformation-resources/profile"
 	"github.com/mongodb/mongodbatlas-cloudformation-resources/util"
 	progressevents "github.com/mongodb/mongodbatlas-cloudformation-resources/util/progressevent"
+	"go.mongodb.org/atlas-sdk/v20230201008/admin"
 )
 
-// Read handles the Read event from the Cloudformation service.
-func Read(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
+func List(req handler.Request, prevModel *Model, currentModel *Model) (handler.ProgressEvent, error) {
 	setup()
-
-	if errEvent := validateModel(ReadRequiredFields, currentModel); errEvent != nil {
+	if errEvent := validateModel(ListRequiredFields, currentModel); errEvent != nil {
 		return *errEvent, nil
 	}
 
@@ -38,28 +36,57 @@ func Read(req handler.Request, prevModel *Model, currentModel *Model) (handler.P
 		currentModel.Profile = aws.String(profile.DefaultProfile)
 	}
 
+	// Create atlas client
 	client, peErr := util.NewAtlasClient(&req, currentModel.Profile)
 	if peErr != nil {
 		return *peErr, nil
 	}
 
-	result, resp, err := client.AtlasV2.ProjectIPAccessListApi.ListProjectIpAccessLists(context.Background(), *currentModel.ProjectId).Execute()
+	pageNum := 0
+	itemsPerPage := 500
+	includeCount := true
+
+	if currentModel.ListOptions != nil {
+		if currentModel.ListOptions.PageNum != nil {
+			pageNum = *currentModel.ListOptions.PageNum
+		}
+
+		if currentModel.ListOptions.IncludeCount != nil {
+			includeCount = *currentModel.ListOptions.IncludeCount
+		}
+
+		if currentModel.ListOptions.ItemsPerPage != nil {
+			itemsPerPage = *currentModel.ListOptions.ItemsPerPage
+		}
+	}
+
+	listOptions := &admin.ListProjectIpAccessListsApiParams{
+		GroupId:      *currentModel.ProjectId,
+		IncludeCount: &includeCount,
+		ItemsPerPage: &itemsPerPage,
+		PageNum:      &pageNum,
+	}
+
+	result, resp, err := client.AtlasV2.ProjectIPAccessListApi.ListProjectIpAccessListsWithParams(context.Background(), listOptions).Execute()
 	if err != nil {
 		return progressevents.GetFailedEventByResponse(fmt.Sprintf("Error getting resource : %s", err.Error()),
 			resp), nil
 	}
 
-	if *result.TotalCount == 0 {
-		return handler.ProgressEvent{
-			Message:          "The entry to read is not in the access list",
-			OperationStatus:  handler.Failed,
-			HandlerErrorCode: cloudformation.HandlerErrorCodeNotFound}, nil
+	mm := make([]AccessListDefinition, 0)
+	for i := range result.Results {
+		var m AccessListDefinition
+		m.completeByConnection(result.Results[i])
+		mm = append(mm, m)
 	}
+	currentModel.AccessList = mm
+	// create list with 1
+	models := []interface{}{}
+	models = append(models, currentModel)
 
-	currentModel.TotalCount = result.TotalCount
 	return handler.ProgressEvent{
 		OperationStatus: handler.Success,
-		Message:         "Read Complete",
-		ResourceModel:   currentModel,
+		Message:         "List Complete",
+		ResourceModels:  models,
 	}, nil
 }
